+++ src/renderer/render.ts
@@ -1,14 +1,16 @@
-﻿import { media } from './media';
+﻿import { media } from './render/media';
+import { math } from './render/math';
 
-export function render(el: Element): void {
+export function render(el: HTMLElement): void {
   assert(el instanceof HTMLImageElement === false);
-  void [el].concat(Array.from(el.querySelectorAll('img:not([src])[data-src], .math')))
+  void [el].concat(Array.from(<NodeListOf<HTMLElement>>el.querySelectorAll('.media, .math')))
     .forEach(el => {
       switch (true) {
-        case el instanceof HTMLImageElement:
+        case el.matches('.media[data-src]'):
+          assert(el instanceof HTMLImageElement);
           return void media(<HTMLImageElement>el);
         case el.matches('.math'):
-          return void MathJax.Hub!.Queue(["Typeset", MathJax.Hub, el]);
+          return void math(el);
         default:
           return;
       }