+++ src/renderer/render/media/pdf.ts
@@ -1,5 +1,5 @@
 ﻿import DOM from 'typed-dom';
-import { parse } from '../parser';
+import { parse } from '../../parser';
 
 export function pdf(url: string): HTMLElement | void {
   if (!url.split(/[?#]/).shift()!.endsWith('.pdf') || url.split('/').length < 4) return;