+++ src/renderer/render/media/slideshare.ts
@@ -1,8 +1,12 @@
 ﻿import DOM from 'typed-dom';
-import { parse } from '../parser';
+import { parse } from '../../parser';
+import { Cache } from 'spica/cache';
+
+const cache = new Cache<string, HTMLElement>(100);
 
 export function slideshare(url: string): HTMLElement | void {
   if (!url.startsWith('https://www.slideshare.net/')) return;
+  if (cache.has(url)) return <HTMLElement>cache.get(url)!.cloneNode(true);
   return DOM.div({
     class: 'media',
     style: 'position: relative;',
@@ -13,6 +17,7 @@ export function slideshare(url: string): HTMLElement | void {
       timeout: 10 * 1e3,
       success({ html }) {
         outer.innerHTML = `<div style="position: relative; padding-top: 83%;">${html}</div>`;
+        void cache.set(url, outer);
         const iframe = outer.querySelector('iframe')!;
         void iframe.setAttribute('style', 'position: absolute; top: 0; bottom: 0; left: 0; right: 0; width: 100%; height: 100%;');
         iframe.parentElement!.style.paddingTop = `${(+iframe.height / +iframe.width) * 100}%`;