+++ src/renderer/render/media/twitter.ts
@@ -1,5 +1,6 @@
 ﻿import DOM from 'typed-dom';
-import { parse } from '../parser';
+import { parse } from '../../parser';
+import { Cache } from 'spica/cache';
 
 declare global {
   interface Window {
@@ -7,8 +8,16 @@ declare global {
   }
 }
 
+let widgetScriptRequested = false;
+const cache = new Cache<string, HTMLElement>(100);
+
 export function twitter(url: string): HTMLElement | void {
   if (!url.startsWith('https://twitter.com/')) return;
+  if (cache.has(url)) {
+    const el = <HTMLElement>cache.get(url)!.cloneNode(true);
+    window.twttr && void window.twttr.widgets.load(el);
+    return el;
+  }
   return DOM.div({
     class: 'media',
     style: 'position: relative;',
@@ -19,7 +28,10 @@ export function twitter(url: string): HTMLElement | void {
       timeout: 10 * 1e3,
       success({ html }): void {
         outer.innerHTML = `<div style="margin-top: -10px; margin-bottom: -10px;">${html}</div>`;
+        void cache.set(url, <HTMLElement>outer.cloneNode(true));
+        if (widgetScriptRequested) return;
         if (window.twttr) return void window.twttr.widgets.load(outer);
+        widgetScriptRequested = true;
         $.ajax(outer.querySelector('script')!.src, {
           dataType: 'script',
           cache: true,